+++ src/main.js
@@ -1,6 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-Vue.mixin({ methods: { t, n } })
 
-const View = Vue.extend(App)
-new View().$mount('#filedrop')
+const app = createApp(App)
+
+app.mixin({ methods: { t, n } })
+
+app.mount('#filedrop')